Fix crash on guild change
When switching from guild A to guild B, Discord checks for permissions of
commands from guild A, even though they are already replaced with ones from
guild B. This was causing a lot of crashes in the hasPermission hook.

This fixes this by simply checking if the application we are checking permission
for actually exists in the guild. This is probably better done in another way,
but hey, it works for now.